# backend/apidocs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for the AffilFlow affiliate marketing platform (Fiber + PostgreSQL).",
        "title": "AffilFlow API",
        "contact": {
            "name": "AffilFlow"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Service root",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.IndexResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payouts/run": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payouts"
                ],
                "summary": "Run affiliate payouts (admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "api"
                ],
                "summary": "Ping API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.PingResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Kubernetes-style health check",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Liveness",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.HealthResponse"
                        }
                    }
                }
            }
        },
        "/ref/{code}": {
            "get": {
                "description": "Records a click, sets attribution cookie, redirects (302)",
                "tags": [
                    "referrals"
                ],
                "summary": "Referral redirect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Referral code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Allowed redirect URL (host must be allowlisted)",
                        "name": "next",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to REDIRECT_BASE_URL or next"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "internal_api_handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.IndexResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string"
                },
                "swagger": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.PingResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "AffilFlow access token: prefix Authorization with Bearer and a space.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}